// src/app/core/interfaces/ClientDTO.ts
export interface ClientDTO {
    clientId?: number;
    names?: string;
    lastNameFat?: string;
    lastNameMot?: string;
    gender?: string;
    telephone?: string;
    email?: string;
    province?: string;
    dateBirth?: string;
    identity?: string;
    direction?: string;
    city?: string;
    postalCode?: string;
    infoClinical?: string;
    notes?: string;
    sendSms?: boolean;
    sendEmail?: boolean;
    active?: boolean;

    // idCliente: string;
    // nombres: string;
    // apellidoPat: string;
    // apellidoMat: string;
    // sexo: string;
    // telefono: string;
    // correo: string;
    // provincia: string;
    // fechaNac: string;
    // dni: string;
    // direccion: string;
    // ciudad: string;
    // codigoPostal: string;
    // infoClinica: string;
    // notas: string;
    // enviosSms: boolean;
    // envioCorreo: boolean;
    // activo: boolean;
}